requirements.txt
google-ads==19.0.0
flask==2.0.1
python-dotenv==0.19.2
pyyaml==6.0
pandas==1.4.2
flask-wtf==1.1.1
requests==2.28.1
requests-oauthlib==1.3.1
python-dateutil==2.8.2
Werkzeug==2.0.3
reportlab==3.6.12
WeasyPrint==54.3
gunicorn==20.1.0
